Common acra-migrate-keys package
Move command-line processing into a shared "migratekeys" package which
could be reused in Acra EE to provide the same command-line interface
as Acra CE without a lot of copypasta.

We'll stil need to use copy-paste inheritance for key store
initialization and the migration process, but that's a necessary evil
because the key store instances are different. Common command-line
parameters are the same, on the other hand.
